#13 Slightly faster implementation Version increment now that Thijs' implementation is the default option

+++ src/RcppExports.cpp
@@ -6,12 +6,12 @@
 using namespace Rcpp;
 
 // get_n_eff_cpp
-std::vector<float> get_n_eff_cpp(const std::vector<double>& z, float comp_width);
+std::vector<float> get_n_eff_cpp(const std::vector<float>& z, float comp_width);
 RcppExport SEXP _comrad_get_n_eff_cpp(SEXP zSEXP, SEXP comp_widthSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const std::vector<double>& >::type z(zSEXP);
+    Rcpp::traits::input_parameter< const std::vector<float>& >::type z(zSEXP);
     Rcpp::traits::input_parameter< float >::type comp_width(comp_widthSEXP);
     rcpp_result_gen = Rcpp::wrap(get_n_eff_cpp(z, comp_width));
     return rcpp_result_gen;
